+++ .github/workflows/build-push-ganache.yml
@@ -1,0 +1,30 @@
+name: Build after merge w/ ganache
+# Builds and runs tests after a PR is merged to main.
+
+on:
+  push:
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Set up Go
+        uses: actions/setup-go@v2
+        with:
+          go-version: 1.17.8
+
+      - name: Setup Node + Ganache
+        uses: actions/setup-node@v3
+        with:
+          node-version: 14
+
+      - name: Setup Ganache
+        run: npm install -g ganache
+
+      - name: Run Ganache
+        run: "ganache --database.dbPath=\"./ganachedb\"  -l 1024000000000 --wallet.accounts=\"0x5dbbff1b5ff19f1ad6ea656433be35f6846e890b3f3ec6ef2b2e2137a8cab4ae,0x56BC75E2D63100000\" --wallet.accounts=\"0xb728cd9a9f54cede03a82fc189eab4830a612703d48b7ef43ceed2cbad1a06c7,0x56BC75E2D63100000\" --wallet.accounts=\"0x1e1e76d5c0ea1382b6acf76e873977fd223c7fa2a6dc57db2b94e93eb303ba85,0x56BC75E2D63100000\" -p 7545 -g 225 --miner.callGasLimit 1024000000000 &"
+
+      - name: Test
+        run: go test -v -tags ganache ./...
